+++ src/pages/login/kakaoSign.tsx
@@ -1,24 +1,14 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { instance } from '@src/api/client';
+import { getLoginToken } from '@src/api/client';
 
 export function KakaoLogin() {
-  const navigate = useNavigate();
   const AUTHORIZE_CODE: string = new URL(window.location.href).searchParams.get(
     'code'
   )!;
 
   useEffect(() => {
-    instance
-      .get(`/users/login?code=${AUTHORIZE_CODE}`)
-      .then((response) => {
-        const JWT = response.data.data.accessToken;
-        if (JWT) localStorage.setItem('jwt', JWT);
-        navigate('/');
-      })
-      .catch(() => {
-        navigate('/login');
-      });
+    getLoginToken(AUTHORIZE_CODE);
   }, []);
 
   return <div>login redirect ...</div>; /* spinner */
@@ -28,8 +18,12 @@ export function KakaoLogout() {
   const navigate = useNavigate();
 
   useEffect(() => {
-    if (localStorage.getItem('jwt')) localStorage.removeItem('jwt');
-    navigate('/login');
+    if (localStorage.getItem('EXIT_LOGIN_TOKEN')) {
+      localStorage.removeItem('EXIT_LOGIN_TOKEN');
+      navigate('/login');
+    } else {
+      navigate('/');
+    }
   }, []);
 
   return <div>logout redirect ...</div>; /* spinner */
